+++ returns/contrib/mypy/returns_plugin.py
@@ -11,9 +11,10 @@ We use ``pytest-mypy-plugins`` to test that it works correctly, see:
 https://github.com/mkurnikov/pytest-mypy-plugins
 """
 
-from typing import Callable, Mapping, Optional, Type
+from typing import Callable, ClassVar, Mapping, Optional, Type
 
-from mypy.plugin import FunctionContext, Plugin
+from mypy.plugin import FunctionContext, MethodContext, MethodSigContext, Plugin
+from mypy.types import CallableType
 from mypy.types import Type as MypyType
 from typing_extensions import Final, final
 
@@ -22,8 +23,12 @@ from returns.contrib.mypy._features import (
     decorators,
     flow,
     partial,
+    pipe,
     pointfree,
 )
+
+# Constant fullnames for typechecking
+# ===================================
 
 #: Set of full names of our decorators.
 _TYPED_DECORATORS: Final = frozenset((
@@ -57,23 +62,51 @@ _TYPED_CURRY_FUNCTION: Final = 'returns.curry.curry'
 #: Used for typed ``flow`` call.
 _TYPED_FLOW_FUNCTION: Final = 'returns._generated.pipeline.flow._flow'
 
+#: Used for typed ``pipe`` call.
+_TYPED_PIPE_FUNCTION: Final = 'returns._generated.pipeline.pipe._pipe'
+_TYPED_PIPE_METHOD: Final = 'returns._generated.pipeline.pipe._Pipe.__call__'
+
+
+# Type aliases
+# ============
+
 #: Type for a function hook.
 _FunctionCallback = Callable[[FunctionContext], MypyType]
+
+#: Type for a method hook.
+_MethodCallback = Callable[[MethodContext], MypyType]
+
+#: Type for a method signature hook.
+_MethodSigCallback = Callable[[MethodSigContext], CallableType]
 
 
 @final
 class _ReturnsPlugin(Plugin):
-    _function_hook_plugins: Mapping[str, _FunctionCallback] = {
+    """Our main plugin to dispatch different callbacks to specific features."""
+
+    _function_hook_plugins: ClassVar[Mapping[str, _FunctionCallback]] = {
         _TYPED_PARTIAL_FUNCTION: partial.analyze,
         _TYPED_CURRY_FUNCTION: curry.analyze,
         _TYPED_FLOW_FUNCTION: flow.analyze,
+        _TYPED_PIPE_FUNCTION: pipe.analyze,
         **dict.fromkeys(_TYPED_POINTFREE_FUNCTIONS, pointfree.analyze),
         **dict.fromkeys(_TYPED_DECORATORS, decorators.analyze),
     }
 
-    def get_function_hook(self, fullname: str) -> Optional[_FunctionCallback]:
+    _method_hook_plugins: ClassVar[Mapping[str, _MethodCallback]] = {
+        _TYPED_PIPE_METHOD: pipe.infer,
+    }
+
+    _method_sig_hook_plugins: ClassVar[Mapping[str, _MethodSigCallback]] = {
+        _TYPED_PIPE_METHOD: pipe.signature,
+    }
+
+    def get_function_hook(
+        self,
+        fullname: str,
+    ) -> Optional[_FunctionCallback]:
         """
-        One of the specified ``mypy`` callbacks.
+        Called for function return types from ``mypy``.
 
         Runs on each function call in the source code.
         We are only interested in a particular subset of all functions.
@@ -82,6 +115,20 @@ class _ReturnsPlugin(Plugin):
         Otherwise, we return ``None``.
         """
         return self._function_hook_plugins.get(fullname)
+
+    def get_method_hook(
+        self,
+        fullname: str,
+    ) -> Optional[_MethodCallback]:
+        """Called for method return types from ``mypy``."""
+        return self._method_hook_plugins.get(fullname)
+
+    def get_method_signature_hook(
+        self,
+        fullname: str,
+    ) -> Optional[_MethodSigCallback]:
+        """Called for method signature from ``mypy``."""
+        return self._method_sig_hook_plugins.get(fullname)
 
 
 def plugin(version: str) -> Type[Plugin]: